Localise JSON types and co-locate them in their respective modules
Make clear in the docs how the top-level `json` module is used.
